setting up cleanup facilities, both pre and post executing downlaod

+++ scripts/get_geocity2.1.php
@@ -12,6 +12,7 @@ if(!defined('NOSESSION')) define('NOSESSION',true);
 
 class qs_geoliteCity {
     private $tempdir;
+	private $helper;
     function __construct () {
             global $conf;
             $this->tempdir = $conf['tmpdir'];
@@ -19,14 +20,21 @@ class qs_geoliteCity {
                 echo "creating tmp directory: " . $this->tempdir . '/tmp' . "\n";
                 mkdir($this->tempdir . '/tmp');
             }
+            else {
+                  echo "Checking for clean working directory\n";
+                  $this->process_gcity(true);
+                  $this->cleanup(true);
+            }
+			$this->helper = plugin_load('helper', 'quickstats');   
+       //     exit;
     }
     
     function get_GeoLiteCity() {
-        @set_time_limit(120);  
-     
-        $helper = plugin_load('helper', 'quickstats');       
+        @set_time_limit(120);     
+       
          $url = "https://geolite.maxmind.com/download/geoip/database/GeoLite2-City.tar.gz";
-     //   $url = "http://epicurus.bz/devel/_media/geolite2-city_20191015.tar.gz";        
+     //   $url = "http://epicurus.bz/devel/_media/geolite2-city_20191015.tar.gz";  
+        $url = "http://epicurus.bz/GeoLite2-City.tar.gz";	 
         $gzfile = $this->tempdir  .  '/GeoLite2-City.tar.gz';    
         
         $http = new DokuHTTPClient();
@@ -36,21 +44,21 @@ class qs_geoliteCity {
 
         $data = $http->get($url);
         if(!$data) { 
-            $this->qs_say($helper->getLang('download_fail'),  $gzfile);
+            $this->qs_say($this->helper->getLang('download_fail'),  $gzfile);
             return;
           }  
 
          $fp = @fopen($gzfile,'wb');
           if($fp === false) { 
-               $this->qs_say($helper->getLang('write_fail'),  $gzfile);
+               $this->qs_say($this->helper->getLang('write_fail'),  $gzfile);
                return;
           }
           if(!fwrite($fp,$data)) {
-             $this->qs_say($helper->getLang('write_fail'),  $gzfile);    
+             $this->qs_say($this->helper->getLang('write_fail'),  $gzfile);    
              return;
           }
           fclose($fp); 
-         $this->qs_say($helper->getLang('file_saved'),  $gzfile);          
+         $this->qs_say($this->helper->getLang('file_saved'),  $gzfile);          
 
         $gz = gzopen($gzfile, "rb");  
         $data= gzread($gz, 32777216);
@@ -60,19 +68,24 @@ class qs_geoliteCity {
 
         
      function qs_unpack() {
-        $ro = ini_get('phar.readonly');
-       // echo $ro . "\n";
+        $ro = ini_get('phar.readonly');       
         if($ro) ini_set('phar.readonly','0');
         $files = scandir($this->tempdir);
       
-        foreach ($files as $file) {
-           //   echo $this->tempdir."/$file\n";
-              if (preg_match("#GeoLite2-City.tar.gz#", $file,$matches)) { 
-              $file = $this->tempdir."/$file";
+        foreach ($files as $file) {          
+            if(preg_match("#GeoLite2-City.tar.gz#", $file,$matches)) { 
+               $file = $this->tempdir."/$file";
                $p = new PharData($file);
+			   $tar = str_replace('.gz', "", $file);
+			   echo "$tar\n";
+			   if(file_exists($tar)){
+				 $this->qs_say($this->helper->getLang('file_exists'),"\n$file\n$tar\n");
+				 exit;
+			   }	   
+			        	   
                $p->decompress(); // creates /path/to/my.tar
-               $tar = str_replace('.gz', "", $file);       
-            //    echo $tar . "\n";
+                    
+            
                 try {
                      $phar = new PharData($tar);
                      $phar->extractTo($this->tempdir.'/tmp'); // extract all files
@@ -83,25 +96,26 @@ class qs_geoliteCity {
        }
     }
      
-  function process_gcity() {   
+  function process_gcity($ini = null) {   
    $tmpdir_files = scandir($this->tempdir . '/' . 'tmp'); 
     foreach ($tmpdir_files as $tmpfile) {
          $current_file = $this->tempdir . '/' . "tmp/${tmpfile}";
           if(preg_match("#(?i)GeoLite2-City_\d+#",$tmpfile) ) {           
               if(is_dir($this->tempdir . '/' . 'tmp/'. $tmpfile)) {                  
-                  $geo_dir_name =  $this->tempdir . '/' . 'tmp/'. $tmpfile;
-               //   echo "$geo_dir_name is a geocity directory\n";
+                  $geo_dir_name =  $this->tempdir . '/' . 'tmp/'. $tmpfile;            
                   $geo_dir = scandir($this->tempdir . '/' . 'tmp/'. $tmpfile);                
-                 // echo "Directory name: $geo_dir_name\n";              
-                  foreach($geo_dir as $gfile) {
+                   foreach($geo_dir as $gfile) {
                       if(!is_dir($this->tempdir . '/' . 'tmp/'. $gfile)) {                          
-                          if(preg_match("/\.mmdb$/",$gfile)) {                           
+                          if(preg_match("/\.mmdb$/",$gfile) && !$ini) {                           
                               echo "renaming " . $geo_dir_name. "/$gfile" ."\nTo: " . MMDB ."\n";
                               rename($geo_dir_name. "/$gfile",MMDB);
                              continue;
                           }    
                            $discard = "$geo_dir_name/$gfile";
-                           echo "Unlinking $discard\n";
+                           if($ini) {
+                               echo "Unlinking $discard\n";    
+                           }                
+                         if(is_writable($discard)) 
                            unlink ($discard);
                       }                     
                   }               
@@ -111,11 +125,15 @@ class qs_geoliteCity {
               }
           }      
       }
-        rmdir($geo_dir_name);
+       if(file_exists($geo_dir_name)) {   
+          rmdir($geo_dir_name);
+       }
+       else echo "Please check your data/temp directory and remove data/tmp/tmp any Geocity files found in it\n";
+        
     }   
     function cleanup() {
         $to_cleanup = scandir($this->tempdir);
-     //   print_r($to_cleanup);
+        //print_r($to_cleanup);
         foreach($to_cleanup as $file) {
             $del = $this->tempdir . '/' . $file;
             if(!is_dir($del) && preg_match("/GeoLite2-City/i",$file)) {
@@ -138,5 +156,5 @@ $geoLite = new qs_geoliteCity();
 $geoLite->get_GeoLiteCity();
 $geoLite->qs_unpack() ;
 $geoLite->process_gcity();
-$geoLite->cleanup();    
+$geoLite->cleanup(true);
